Fix templ var conflict on: videos,audios,subtitles,chapters
It renames them to: videos_pretty, audio_pretty, subtitles_pretty, and chapters_yes_no.

This allows the underlying (already existing) videos, audio, subtitles, and chapters to be usable as well.

This allows the len: filters on the videos, audio, subtitles, and chapters to be the correct length (and not think the `--` from *_pretty is a valid item to count).