NFO: Optimize readability and usefulness of attribute inits
This removes the real initialization of the variable (gets rid of the = None) because once real type hinting is made, it from then on thinks it can also be None, which is in no way ideal.

The type hinting done here does not actually initialize anything but allows readers to see what attributes it has and what they can be, a lot easier and better than before.