generate: Allow template args to override or be overridden by config
E.g. let's say you set the config to override the Source variable, but the template args also set the config variable, then it would crash and panic on which one to use as it was re-declared.

Or, let's say source variable is defined by the nfo generate arguments, and the template args wants to override it, it now can.

Priority order is Config > Template Args > Generator Args