generate: Don't accept config values for the NFO config args
This could be abused unknowingly if the template tries to get values from the config on purpose, e.g. `{generate.art}` or `{fanart_api_key}` if done so.

It also generally has no need to set values with the config, there's no reason to. Anything that can be done statically like such, should be written directly within the template (or art) itself.