Merge get_imdb_id, get_tmdb_id, and get_tvdb_id code into NFO constructor
The user-input loop was removed entirely, as well as the warnings for no TMDB or TVDB IDs.

The input loop was removed as we can assume the user could just retry the CLI call to update the IDs. The warnings were removed as the user would actively have to go out of their way not to put the IDs, they know they didn't put them.